test/e2e/framework:move functions to test/e2e/scheduling/
